Move exception classes into exception package
