+++ src/main/kotlin/io/github/janbarari/gradle/analytics/exception/NotAccessibleGitTerminalException.kt
@@ -20,12 +20,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.janbarari.gradle.analytics
+package io.github.janbarari.gradle.analytics.exception
 
 import io.github.janbarari.gradle.ExcludeJacocoGenerated
-import java.io.File
+import io.github.janbarari.gradle.analytics.GradleAnalyticsPlugin
 
 @ExcludeJacocoGenerated
-class PluginConfigNotValidException(msg: String, buildScript: File): Throwable(
-    message = "${GradleAnalyticsPlugin.PLUGIN_NAME}: $msg\n${buildScript.absolutePath}"
+class NotAccessibleGitTerminalException: Throwable(
+    message = "${GradleAnalyticsPlugin.PLUGIN_NAME} works only on projects that use Git."
 )
